Add files via upload
This Dashboard is mainly constrained of a global sales dataset where we first have connected the database to the powerbi desktop to analysis the data after which data cleaning was performed to remove the null values from the columns and bridge the gap of missing values to get the most accurate results for better decisions.  By the end of this project data model and data visualization was performed to create an interactive dashboard for better understanding.